+++ wazo_auth/plugins/http/password_reset/http.py
@@ -1,4 +1,4 @@
-# Copyright 2018-2021 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2018-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
@@ -29,9 +29,10 @@ class PasswordReset(http.ErrorCatchingResource):
             raise PasswordResetException.from_errors(e.messages)
 
         logger.debug(
-            'resetting password for %s', args['username'] or args['email_address']
+            'resetting password for %s',
+            args['username'] or args['email_address'] or args['login'],
         )
-        search_params = {k: v for k, v in args.items() if v}
+        search_params = {field: value for field, value in args.items() if value}
         users = self.user_service.list_users(**search_params)
         if not users:
             # We do not want to leak the information if a user exists or not