+++ wazo_auth/plugins/http/password_reset/tests/test_schemas.py
@@ -31,28 +31,43 @@ class TestSchema(unittest.TestCase):
 
         assert_that(result, has_entries(email_address=email))
 
-    def test_that_username_and_email_are_mutually_exclusive(self):
-        query_string = {'email': 'foo@bar.com', 'username': 'foobar'}
+    def test_that_username_and_email_and_login_are_mutually_exclusive(self):
+        queries = [
+            {'email': 'foo@bar.com', 'username': 'foobar', 'login': 'foo'},
+            {'email': 'foo@bar.com', 'username': None, 'login': 'foo'},
+            {'email': 'foo@bar.com', 'username': 'foobar', 'login': None},
+            {'email': None, 'username': 'foobar', 'login': 'foo'},
+        ]
 
-        assert_that(
-            calling(self.password_query_parameters_schema.load).with_args(query_string),
-            raises(
-                ValidationError,
-                has_property(
-                    "messages",
-                    has_entries(
-                        _schema=contains_exactly('"username" or "email" should be used')
+        for query in queries:
+            assert_that(
+                calling(self.password_query_parameters_schema.load).with_args(query),
+                raises(
+                    ValidationError,
+                    has_property(
+                        "messages",
+                        has_entries(
+                            _schema=contains_exactly(
+                                '"username" or "email" or "login" should be used'
+                            )
+                        ),
                     ),
                 ),
-            ),
-        )
+            )
 
     def test_username_only(self):
         query_string = {'username': 'foobar'}
 
         result = self.password_query_parameters_schema.load(query_string)
 
-        assert_that(result, has_entries(username='foobar', email_address=None))
+        assert_that(
+            result,
+            has_entries(
+                username='foobar',
+                email_address=None,
+                login=None,
+            ),
+        )
 
     def test_email_only(self):
         query_string = {'email': 'foobar@example.com'}
@@ -60,7 +75,26 @@ class TestSchema(unittest.TestCase):
         result = self.password_query_parameters_schema.load(query_string)
 
         assert_that(
-            result, has_entries(username=None, email_address='foobar@example.com')
+            result,
+            has_entries(
+                username=None,
+                email_address='foobar@example.com',
+                login=None,
+            ),
+        )
+
+    def test_login_only(self):
+        query_string = {'login': 'foobar@example.com'}
+
+        result = self.password_query_parameters_schema.load(query_string)
+
+        assert_that(
+            result,
+            has_entries(
+                username=None,
+                email_address=None,
+                login='foobar@example.com',
+            ),
         )
 
     def test_invalid_field(self):